add global conversion method for ms to date to be used by all components
